make new helper an extension method
